Preserve menu bar, tool bar, and tab bar
When setting a new global map, the entries for the menu bar, tool bar, and
tab bar must be copied to the new map from the old one to avoid hiding the
bars.

Before this change, evil-owl's popup would hide these bars when it was
opened, and the tab bar was not restored when the popup was closed.  After
this change, none of the bars are hidden.

* evil-owl.el (evil-owl--with-popup-map): Copy the old global map's
menu-bar, tool-bar, and tab-bar entries to the new map.